fix: Create getInformation in utils to make the code cleaner in Text
